requirements.txt
python-dateutil==1.5
snakebite==2.9.0
hdfs==2.0.7
